+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ### ADDITIONS:
+- Added Decoration Blocks
 
 ### FIXES:
-- fixed space crashing when using upgraded non-spacesuit armor
 
 ### CHANGES: